[ISSUE-415][REFACTOR] Refactor Storage related class to separated scala file (#416)
