+++ server-worker/src/main/java/com/aliyun/emr/rss/service/deploy/worker/storage/FileWriter.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package com.aliyun.emr.rss.service.deploy.worker;
+package com.aliyun.emr.rss.service.deploy.worker.storage;
 
 import java.io.File;
 import java.io.FileOutputStream;
@@ -25,7 +25,6 @@ import java.util.List;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.concurrent.atomic.AtomicReference;
 
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.CompositeByteBuf;
@@ -42,6 +41,7 @@ import com.aliyun.emr.rss.common.network.server.MemoryTracker;
 import com.aliyun.emr.rss.common.protocol.PartitionSplitMode;
 import com.aliyun.emr.rss.common.protocol.PartitionType;
 import com.aliyun.emr.rss.common.protocol.StorageInfo;
+import com.aliyun.emr.rss.service.deploy.worker.WorkerSource;
 
 /*
  * Note: Once FlushNotifier.exception is set, the whole file is not available.
@@ -89,37 +89,17 @@ public final class FileWriter implements DeviceObserver {
     deviceMonitor.unregisterFileWriter(this);
   }
 
-  static class FlushNotifier {
-    final AtomicInteger numPendingFlushes = new AtomicInteger();
-    final AtomicReference<IOException> exception = new AtomicReference<>();
-
-    void setException(IOException e) {
-      exception.set(e);
-    }
-
-    boolean hasException() {
-      return exception.get() != null;
-    }
-
-    void checkException() throws IOException {
-      IOException e = exception.get();
-      if (e != null) {
-        throw e;
-      }
-    }
-  }
-
   private final FlushNotifier notifier = new FlushNotifier();
 
   public FileWriter(
-    FileInfo fileInfo,
-    Flusher flusher,
-    AbstractSource workerSource,
-    RssConf rssConf,
-    DeviceMonitor deviceMonitor,
-    long splitThreshold,
-    PartitionSplitMode splitMode,
-    PartitionType partitionType) throws IOException {
+      FileInfo fileInfo,
+      Flusher flusher,
+      AbstractSource workerSource,
+      RssConf rssConf,
+      DeviceMonitor deviceMonitor,
+      long splitThreshold,
+      PartitionSplitMode splitMode,
+      PartitionType partitionType) throws IOException {
     this.fileInfo = fileInfo;
     this.flusher = flusher;
     this.flushWorkerIndex = flusher.getWorkerIndex();
