+++ server-worker/src/main/scala/com/aliyun/emr/rss/service/deploy/worker/storage/DeviceObserver.scala
@@ -15,22 +15,12 @@
  * limitations under the License.
  */
 
-package com.aliyun.emr.rss.service.deploy.worker;
+package com.aliyun.emr.rss.service.deploy.worker.storage
 
-import com.aliyun.emr.rss.common.protocol.PartitionLocation;
-import com.aliyun.emr.rss.service.deploy.worker.storage.FileWriter;
+import com.aliyun.emr.rss.common.meta.DiskStatus
 
-public class WorkingPartition extends PartitionLocation {
-  private final transient FileWriter fileWriter;
-
-  public WorkingPartition(
-      PartitionLocation partitionLocation,
-      FileWriter fileWriter) {
-    super(partitionLocation);
-    this.fileWriter = fileWriter;
-  }
-
-  public FileWriter getFileWriter() {
-    return fileWriter;
-  }
+trait DeviceObserver {
+  def notifyError(mountPoint: String, diskStatus: DiskStatus): Unit = {}
+  def notifyHealthy(mountPoint: String): Unit = {}
+  def notifyHighDiskUsage(mountPoint: String): Unit = {}
 }
